KAFKA-14705: Move topic filters to tools module
After MirrorMaker 1 removal, there are no other modules dependencies for these classes, so we can safely move them to tools module.
