+++ tools/src/main/java/org/apache/kafka/tools/GetOffsetShell.java
@@ -31,14 +31,14 @@ import org.apache.kafka.common.utils.Exit;
 import org.apache.kafka.common.utils.Utils;
 import org.apache.kafka.server.util.CommandDefaultOptions;
 import org.apache.kafka.server.util.CommandLineUtils;
-import org.apache.kafka.server.util.PartitionFilter;
-import org.apache.kafka.server.util.PartitionFilter.PartitionRangeFilter;
-import org.apache.kafka.server.util.PartitionFilter.PartitionsSetFilter;
-import org.apache.kafka.server.util.PartitionFilter.UniquePartitionFilter;
-import org.apache.kafka.server.util.TopicFilter.IncludeList;
-import org.apache.kafka.server.util.TopicPartitionFilter;
-import org.apache.kafka.server.util.TopicPartitionFilter.CompositeTopicPartitionFilter;
-import org.apache.kafka.server.util.TopicPartitionFilter.TopicFilterAndPartitionFilter;
+import org.apache.kafka.tools.filter.PartitionFilter;
+import org.apache.kafka.tools.filter.PartitionFilter.PartitionRangeFilter;
+import org.apache.kafka.tools.filter.PartitionFilter.PartitionsSetFilter;
+import org.apache.kafka.tools.filter.PartitionFilter.UniquePartitionFilter;
+import org.apache.kafka.tools.filter.TopicFilter.IncludeList;
+import org.apache.kafka.tools.filter.TopicPartitionFilter;
+import org.apache.kafka.tools.filter.TopicPartitionFilter.CompositeTopicPartitionFilter;
+import org.apache.kafka.tools.filter.TopicPartitionFilter.TopicFilterAndPartitionFilter;
 
 import java.io.IOException;
 import java.util.Arrays;
