+++ tools/src/main/java/org/apache/kafka/tools/ReplicaVerificationTool.java
@@ -55,7 +55,7 @@ import org.apache.kafka.common.utils.Utils;
 import org.apache.kafka.server.util.CommandDefaultOptions;
 import org.apache.kafka.server.util.CommandLineUtils;
 import org.apache.kafka.server.util.ShutdownableThread;
-import org.apache.kafka.server.util.TopicFilter;
+import org.apache.kafka.tools.filter.TopicFilter;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
