+++ tools/src/main/java/org/apache/kafka/tools/TopicCommand.java
@@ -50,7 +50,7 @@ import org.apache.kafka.server.common.AdminCommandFailedException;
 import org.apache.kafka.server.common.AdminOperationException;
 import org.apache.kafka.server.util.CommandDefaultOptions;
 import org.apache.kafka.server.util.CommandLineUtils;
-import org.apache.kafka.server.util.TopicFilter.IncludeList;
+import org.apache.kafka.tools.filter.TopicFilter.IncludeList;
 import org.apache.kafka.storage.internals.log.LogConfig;
 
 import org.slf4j.Logger;
