+++ tools/src/main/java/org/apache/kafka/tools/filter/TopicFilter.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.server.util;
+package org.apache.kafka.tools.filter;
 
 import org.apache.kafka.common.internals.Topic;
